🔨 drop global no-data pattern
