+++ packages/@ourworldindata/grapher/src/mapCharts/ChoroplethGlobe.tsx
@@ -55,9 +55,9 @@ import {
     CountryWithNoData,
     ExternalValueAnnotation,
     InternalValueAnnotation,
-    NoDataPattern,
     ProjectedDataPattern,
 } from "./MapComponents"
+import { NoDataPattern } from "../chart/ChartUtils"
 import { Patterns } from "../core/GrapherConstants"
 import {
     calculateDistance,
